utils/dataset_manifest/requirements.txt
opencv-python-headless==4.4.0.42
Pillow==7.2.0